requirements.txt
browser-use
langchain-openai
python-dotenv
pydantic